+++ src/Components/OrderCreate/OrderCreateView.js
@@ -8,7 +8,11 @@ export const OrderCreateView = (props) => {
   const { order } = usePelcro();
   const skipPayment =
     window.Pelcro?.uiSettings?.skipPaymentForFreePlans;
-  const showOrderButton = skipPayment && order?.price === 0;
+  const showOrderButton =
+    skipPayment &&
+    (order?.price === 0 ||
+      (order?.length > 0 &&
+        order.every((item) => item?.price === 0)));
 
   return (
     <div id="pelcro-order-create-view">
@@ -21,6 +25,7 @@ export const OrderCreateView = (props) => {
           showCoupon={true}
           showExternalPaymentMethods={false}
           showOrderButton={showOrderButton}
+          order={order}
           {...props}
         />
       </form>