+++ src/Components/PaymentMethod/PaymentMethodContainer.js
@@ -118,6 +118,9 @@ const PaymentMethodContainerWithoutStripe = ({
   onFailure = () => {},
   ...props
 }) => {
+  const [vantivPaymentRequest, setVantivPaymentRequest] =
+    useState(null);
+  const [updatedCouponCode, setUpdatedCouponCode] = useState("");
   const { t } = useTranslation("payment");
   const pelcroStore = usePelcro();
   const { set, order, selectedPaymentMethodId, couponCode } =
@@ -158,7 +161,7 @@ const PaymentMethodContainerWithoutStripe = ({
   }, []);
 
   /*====== Start Tap integration ========*/
-  const submitUsingTap = () => {
+  const submitUsingTap = (state) => {
     const isUsingExistingPaymentMethod = Boolean(
       selectedPaymentMethodId
     );
@@ -278,7 +281,7 @@ const PaymentMethodContainerWithoutStripe = ({
                         }
                       });
 
-                      handleTapPayment(tapID);
+                      handleTapPayment(tapID, state);
                     }
                   };
 
@@ -297,7 +300,7 @@ const PaymentMethodContainerWithoutStripe = ({
       });
   };
 
-  function handleTapPayment(paymentRequest) {
+  function handleTapPayment(paymentRequest, state) {
     const isUsingExistingPaymentMethod = Boolean(
       selectedPaymentMethodId
     );
@@ -518,7 +521,7 @@ const PaymentMethodContainerWithoutStripe = ({
   };
   /*====== End Tap integration ========*/
 
-  const submitUsingVantiv = () => {
+  const submitUsingVantiv = (state) => {
     const isUsingExistingPaymentMethod = Boolean(
       selectedPaymentMethodId
     );
@@ -543,7 +546,7 @@ const PaymentMethodContainerWithoutStripe = ({
     });
   };
 
-  function handleVantivPayment(paymentRequest) {
+  function handleVantivPayment(paymentRequest, couponCode) {
     if (paymentRequest) {
       const SUCCESS_STATUS = "870";
       if (paymentRequest.response !== SUCCESS_STATUS) {
@@ -628,9 +631,6 @@ const PaymentMethodContainerWithoutStripe = ({
       const giftSubscriprition = isGift && !subscriptionIdToRenew;
       const renewGift = isRenewingGift;
 
-      const couponCode =
-        state.couponCode || window.Pelcro.coupon.getFromUrl() || "";
-
       if (renewGift) {
         return payment.execute(
           {
@@ -714,6 +714,7 @@ const PaymentMethodContainerWithoutStripe = ({
             if (err) {
               return handlePaymentError(err);
             }
+
             onSuccess(res);
           }
         );
@@ -726,7 +727,7 @@ const PaymentMethodContainerWithoutStripe = ({
   const tapInstanceCard = React.useRef(null);
 
   useEffect(() => {
-    if (skipPayment && (plan?.amount === 0 || order?.price === 0)) return;
+    if (skipPayment && (plan?.amount === 0 || props?.freeOrders)) return;
     if (cardProcessor === "vantiv" && !selectedPaymentMethodId) {
       const payPageId =
         window.Pelcro.site.read()?.vantiv_gateway_settings
@@ -742,7 +743,8 @@ const PaymentMethodContainerWithoutStripe = ({
         height: "245",
         timeout: 50000,
         div: "eProtectiframe",
-        callback: handleVantivPayment,
+        callback: (paymentRequest) =>
+          setVantivPaymentRequest(paymentRequest),
         showCvv: true,
         numYears: 8,
         placeholderText: {
@@ -758,9 +760,16 @@ const PaymentMethodContainerWithoutStripe = ({
     }
   }, [selectedPaymentMethodId]);
 
+  //Trigger the handleVantivPayment method when a vantivePaymentRequest is present
+  useEffect(() => {
+    if (vantivPaymentRequest) {
+      handleVantivPayment(vantivPaymentRequest, updatedCouponCode);
+    }
+  }, [vantivPaymentRequest]);
+
   useEffect(() => {
     whenUserReady(() => {
-      if (skipPayment && (plan?.amount === 0 || order?.price === 0)) return;
+      if (skipPayment && (plan?.amount === 0 || props?.freeOrders)) return;
       if (cardProcessor === "tap" && !window.Tapjsli) {
         window.Pelcro.helpers.loadSDK(
           "https://cdnjs.cloudflare.com/ajax/libs/bluebird/3.3.4/bluebird.min.js",
@@ -792,7 +801,7 @@ const PaymentMethodContainerWithoutStripe = ({
   }, [selectedPaymentMethodId]);
 
   const initPaymentRequest = (state, dispatch) => {
-    if (skipPayment && (plan?.amount === 0 || order?.price === 0)) return;
+    if (skipPayment && (plan?.amount === 0 || props?.freeOrders)) return;
     try {
       const paymentRequest = stripe.paymentRequest({
         country: window.Pelcro.user.location.countryCode || "US",
@@ -847,7 +856,7 @@ const PaymentMethodContainerWithoutStripe = ({
    * Updates the total amount after adding taxes only if site taxes are enabled
    */
   const updateTotalAmountWithTax = () => {
-    if (skipPayment && (plan?.amount === 0 || order?.price === 0)) return;
+    if (skipPayment && (plan?.amount === 0 || props?.freeOrders)) return;
     const taxesEnabled = window.Pelcro.site.read()?.taxes_enabled;
 
     if (taxesEnabled && type === "createPayment") {
@@ -886,6 +895,9 @@ const PaymentMethodContainerWithoutStripe = ({
 
       if (err) {
         onFailure(err);
+
+        //reset the coupon code in local state
+        setUpdatedCouponCode("");
 
         dispatch({
           type: SET_COUPON_ERROR,
@@ -943,6 +955,9 @@ const PaymentMethodContainerWithoutStripe = ({
         type: SET_COUPON,
         payload: res.data.coupon
       });
+
+      //set the coupon code in local state to be able to use with Vantiv
+      setUpdatedCouponCode(res.data.coupon.code);
 
       dispatch({
         type: SET_PERCENT_OFF,
@@ -1028,6 +1043,9 @@ const PaymentMethodContainerWithoutStripe = ({
 
   const removeAppliedCoupon = (state) => {
     state.couponCode = "";
+
+    //reset the coupon code in local state
+    setUpdatedCouponCode("");
 
     dispatch({ type: SET_COUPON_ERROR, payload: "" });
 
@@ -1469,7 +1487,7 @@ const PaymentMethodContainerWithoutStripe = ({
   };
 
   const submitPayment = (state, dispatch) => {
-    if (skipPayment && order?.price === 0) {
+    if (skipPayment && props?.freeOrders) {
       const isQuickPurchase = !Array.isArray(order);
       const mappedOrderItems = isQuickPurchase
         ? [{ sku_id: order.id, quantity: order.quantity }]
@@ -1777,8 +1795,12 @@ const PaymentMethodContainerWithoutStripe = ({
           return UpdateWithSideEffect(
             { ...state, disableSubmit: true, isLoading: true },
             (state, dispatch) => {
+              if (skipPayment && props?.freeOrders) {
+                return submitPayment(state, dispatch);
+              }
+              
               if (getSiteCardProcessor() === "vantiv") {
-                return submitUsingVantiv();
+                return submitUsingVantiv(state);
               }
 
               if (getSiteCardProcessor() === "tap") {
@@ -1964,7 +1986,7 @@ const PaymentMethodContainer = (props) => {
         </Elements>
       </StripeProvider>
     );
-  } else {
+  } else if (cardProcessor !== "stripe") {
     return (
       <PaymentMethodContainerWithoutStripe store={store} {...props} />
     );